+++ junos/resource_system_login_user.go
@@ -55,8 +55,9 @@ func resourceSystemLoginUser() *schema.Resource {
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"encrypted_password": {
-							Type:     schema.TypeString,
-							Optional: true,
+							Type:          schema.TypeString,
+							Optional:      true,
+							ConflictsWith: []string{"authentication.0.plain_text_password"},
 						},
 						"no_public_keys": {
 							Type:     schema.TypeBool,
@@ -64,6 +65,12 @@ func resourceSystemLoginUser() *schema.Resource {
 							ConflictsWith: []string{
 								"authentication.0.ssh_public_keys",
 							},
+						},
+						"plain_text_password": {
+							Type:          schema.TypeString,
+							Optional:      true,
+							Sensitive:     true,
+							ConflictsWith: []string{"authentication.0.encrypted_password"},
 						},
 						"ssh_public_keys": {
 							Type:          schema.TypeSet,
@@ -154,8 +161,9 @@ func resourceSystemLoginUserRead(ctx context.Context, d *schema.ResourceData, m 
 
 func resourceSystemLoginUserReadWJnprSess(
 	d *schema.ResourceData, m interface{}, jnprSess *NetconfObject) diag.Diagnostics {
+	plainTextPassword := readSystemLoginUserReadDataPlainTextPassword(d)
 	mutex.Lock()
-	systemLoginUserOptions, err := readSystemLoginUser(d.Get("name").(string), m, jnprSess)
+	systemLoginUserOptions, err := readSystemLoginUser(d.Get("name").(string), plainTextPassword, m, jnprSess)
 	mutex.Unlock()
 	if err != nil {
 		return diag.FromErr(err)
@@ -261,7 +269,7 @@ func resourceSystemLoginUserImport(d *schema.ResourceData, m interface{}) ([]*sc
 	if !systemLoginUserExists {
 		return nil, fmt.Errorf("don't find system login user with id '%v' (id must be <name>)", d.Id())
 	}
-	systemLoginUserOptions, err := readSystemLoginUser(d.Id(), m, jnprSess)
+	systemLoginUserOptions, err := readSystemLoginUser(d.Id(), "", m, jnprSess)
 	if err != nil {
 		return nil, err
 	}
@@ -292,49 +300,65 @@ func setSystemLoginUser(d *schema.ResourceData, m interface{}, jnprSess *Netconf
 
 	configSet = append(configSet, setPrefix+"class "+d.Get("class").(string))
 
-	if d.Get("uid").(int) != 0 {
-		configSet = append(configSet, setPrefix+"uid "+strconv.Itoa(d.Get("uid").(int)))
+	if v := d.Get("uid").(int); v != 0 {
+		configSet = append(configSet, setPrefix+"uid "+strconv.Itoa(v))
 	}
-	for _, v := range d.Get("authentication").([]interface{}) {
-		if v == nil {
+	for _, block := range d.Get("authentication").([]interface{}) {
+		if block == nil {
 			return fmt.Errorf("authentication block is empty")
 		}
-		authentication := v.(map[string]interface{})
-		if authentication["encrypted_password"].(string) != "" {
-			configSet = append(configSet, setPrefix+"authentication encrypted-password \""+
-				authentication["encrypted_password"].(string)+"\"")
+		authentication := block.(map[string]interface{})
+		if pass := authentication["plain_text_password"].(string); pass != "" {
+			configSet = append(configSet, setPrefix+"authentication plain-text-password-value \""+pass+"\"")
+		}
+		if pass := authentication["encrypted_password"].(string); pass != "" {
+			configSet = append(configSet, setPrefix+"authentication encrypted-password \""+pass+"\"")
 		}
 		if authentication["no_public_keys"].(bool) {
 			configSet = append(configSet, setPrefix+"authentication no-public-keys")
 		}
-		for _, v2 := range sortSetOfString(authentication["ssh_public_keys"].(*schema.Set).List()) {
+		for _, key := range sortSetOfString(authentication["ssh_public_keys"].(*schema.Set).List()) {
 			switch {
-			case strings.HasPrefix(v2, ssh.KeyAlgoDSA):
-				configSet = append(configSet, setPrefix+"authentication ssh-dsa \""+v2+"\"")
-			case strings.HasPrefix(v2, ssh.KeyAlgoRSA):
-				configSet = append(configSet, setPrefix+"authentication ssh-rsa \""+v2+"\"")
-			case strings.HasPrefix(v2, ssh.KeyAlgoECDSA256),
-				strings.HasPrefix(v2, ssh.KeyAlgoECDSA384),
-				strings.HasPrefix(v2, ssh.KeyAlgoECDSA521):
-				configSet = append(configSet, setPrefix+"authentication ssh-ecdsa \""+v2+"\"")
-			case strings.HasPrefix(v2, ssh.KeyAlgoED25519):
-				configSet = append(configSet, setPrefix+"authentication ssh-ed25519 \""+v2+"\"")
+			case strings.HasPrefix(key, ssh.KeyAlgoDSA):
+				configSet = append(configSet, setPrefix+"authentication ssh-dsa \""+key+"\"")
+			case strings.HasPrefix(key, ssh.KeyAlgoRSA):
+				configSet = append(configSet, setPrefix+"authentication ssh-rsa \""+key+"\"")
+			case strings.HasPrefix(key, ssh.KeyAlgoECDSA256),
+				strings.HasPrefix(key, ssh.KeyAlgoECDSA384),
+				strings.HasPrefix(key, ssh.KeyAlgoECDSA521):
+				configSet = append(configSet, setPrefix+"authentication ssh-ecdsa \""+key+"\"")
+			case strings.HasPrefix(key, ssh.KeyAlgoED25519):
+				configSet = append(configSet, setPrefix+"authentication ssh-ed25519 \""+key+"\"")
 			default:
-				return fmt.Errorf("format in public key '%v' not supported", v2)
+				return fmt.Errorf("format in public key '%v' not supported", key)
 			}
 		}
 	}
-	if d.Get("cli_prompt").(string) != "" {
-		configSet = append(configSet, setPrefix+"cli prompt \""+d.Get("cli_prompt").(string)+"\"")
+	if v := d.Get("cli_prompt").(string); v != "" {
+		configSet = append(configSet, setPrefix+"cli prompt \""+v+"\"")
 	}
-	if d.Get("full_name").(string) != "" {
-		configSet = append(configSet, setPrefix+"full-name \""+d.Get("full_name").(string)+"\"")
+	if v := d.Get("full_name").(string); v != "" {
+		configSet = append(configSet, setPrefix+"full-name \""+v+"\"")
 	}
 
 	return sess.configSet(configSet, jnprSess)
 }
 
-func readSystemLoginUser(name string, m interface{}, jnprSess *NetconfObject) (systemLoginUserOptions, error) {
+func readSystemLoginUserReadDataPlainTextPassword(d *schema.ResourceData) string {
+	if blocks := d.Get("authentication").([]interface{}); len(blocks) > 0 {
+		if blocks[0] == nil {
+			return ""
+		}
+		auth := blocks[0].(map[string]interface{})
+
+		return auth["plain_text_password"].(string)
+	}
+
+	return ""
+}
+
+func readSystemLoginUser(name, plainTextPassword string, m interface{}, jnprSess *NetconfObject,
+) (systemLoginUserOptions, error) {
 	sess := m.(*Session)
 	var confRead systemLoginUserOptions
 
@@ -364,15 +388,20 @@ func readSystemLoginUser(name string, m interface{}, jnprSess *NetconfObject) (s
 			case strings.HasPrefix(itemTrim, "authentication "):
 				if len(confRead.authentication) == 0 {
 					confRead.authentication = append(confRead.authentication, map[string]interface{}{
-						"encrypted_password": "",
-						"no_public_keys":     false,
-						"ssh_public_keys":    make([]string, 0),
+						"encrypted_password":  "",
+						"no_public_keys":      false,
+						"plain_text_password": "",
+						"ssh_public_keys":     make([]string, 0),
 					})
 				}
 				switch {
 				case strings.HasPrefix(itemTrim, "authentication encrypted-password "):
-					confRead.authentication[0]["encrypted_password"] = strings.Trim(strings.TrimPrefix(
-						itemTrim, "authentication encrypted-password "), "\"")
+					if plainTextPassword != "" {
+						confRead.authentication[0]["plain_text_password"] = plainTextPassword
+					} else {
+						confRead.authentication[0]["encrypted_password"] = strings.Trim(strings.TrimPrefix(
+							itemTrim, "authentication encrypted-password "), "\"")
+					}
 				case itemTrim == "authentication no-public-keys":
 					confRead.authentication[0]["no_public_keys"] = true
 				case strings.HasPrefix(itemTrim, "authentication ssh-dsa "):