+++ junos/resource_system_login_user_test.go
@@ -63,6 +63,13 @@ resource "junos_system_login_user" "testacc2" {
   name  = "test.acc2"
   class = "unauthorized"
 }
+resource "junos_system_login_user" "testacc3" {
+  name  = "test.acc3"
+  class = "unauthorized"
+  authentication {
+    plain_text_password = "test1234"
+  }
+}
 `
 }